functions/process-upload/requirements.txt
google-cloud-storage
filetype

# sqlite utils
csvs-to-sqlite
yaml-to-sqlite
markdown-to-sqlite